# swagger-codegen-demo/src/main/java/com/example/swaggercodegendemo/StudentController.java
package com.example.swaggercodegendemo;

import java.net.URI;
import java.util.List;
import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import io.swagger.annotations.ApiOperation;


@RestController
public class StudentController {

    @Autowired
    private StudentRepository studentRepository;

    @GetMapping("/students")
    public List<Student> getAllStudents() {
        return studentRepository.findAll();
    }

    @GetMapping("/students/{id}")
    public Student getStudent(@PathVariable long id) {
    	
        Optional<Student> student = studentRepository.findById(id);

        if (!student.isEmpty()) {
        	return student.get();
        }
        
        throw new StudentNotFoundException("id: " + id);        
    }

    @ApiOperation(value = "Add a Student",
    	    notes = "Also returns a link to the new added Student")
    @PostMapping("/students")
    public ResponseEntity<Object> createStudent(@RequestBody Student student) {
    	
    	Student savedStudent = studentRepository.save(student);

        URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
                .buildAndExpand(savedStudent.getId()).toUri();

        return ResponseEntity.created(location).build();

    }

    @PutMapping("/students/{id}")
    public ResponseEntity<Object> updateStudent(@RequestBody Student student, @PathVariable long id) {

        Optional<Student> studentOptional = studentRepository.findById(id);

        if (studentOptional.isEmpty()) {
        	return ResponseEntity.notFound().build();
        }            

        student.setId(id);

        studentRepository.save(student);

        return ResponseEntity.noContent().build();
    }
    
    @DeleteMapping("/students/{id}")
    public void deleteStudent(@PathVariable long id) {
        studentRepository.deleteById(id);
    }
    
}